requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.0